// web-app/src/app/results/interfaces/interfaces.ts
import Image from "next/image";

export interface UserDisplayData {
  country: string;
  display_name: string;
  email: string;
  explicit_content?: ExplicitContent;
  external_urls?: ExternalUrls;
  followers?: Followers;
  href: string;
  id: string;
  images: Image[];
  product: string;
  type: string;
  uri: string;
}

export interface ExplicitContent {
  filter_enabled: boolean;
  filter_locked: boolean;
}

export interface ExternalUrls {
  spotify: string;
}

export interface Followers {
  href: string;
  total: number;
}

export interface Image {
  url: string;
  height: number;
  width: number;
}

export interface TrackData {
  UID: string;
  trackName: string;
  artistName: string;
  total_msPlayed: number;
}

export interface ArtistData {
  UID: string;
  artistName: string;
  total_msPlayed: number;
}

export interface TotalPlayTime {
  UID: string;
  total_msPlayed: number;
}

export interface JSONResponseData {
  spotify_uid: string;
  top_songs: TrackData[];
  top_artist: ArtistData[];
  total_ms_played: TotalPlayTime[];
}